fix: attenddes list
