+++ apps/littleape/components/vite-migrated/index.ts
@@ -1,0 +1,2 @@
+export { default as Attendees } from './Attendees';
+export { default as Participant } from './Participant'
